Move raster IO operations into op.io package.
