+++ docs/00.md
@@ -21,6 +21,10 @@ operations with existing operations.
 ## Some sample GeoTrellis code
 
 ```scala
+
+  // Import a few raster operations
+  import op.raster.local.{Add,MultiplyConstant,DivideConstant}
+
   // create a new operation that multiplies each cell of 
   // each raster by a weight, and then add those two new
   // rasters together
