+++ docs/02.md
@@ -15,102 +15,5 @@ included in this documentation to start exploring how to build your own geoproce
 you won't need to install Scala or any other files to get started.  The following 
 instructions assume you are running under Linux or OS X.
 
-## Step 1: Run SBT
-
-This project uses SBT, Simple Build Tool, for compilation and execution.
-
-If you don't have SBT installed already, you can run a script in the template
-directory (from https://github.com/paulp/sbt-extras) that will automatically download it for you.
-
-At the command line, go to the directory where you've installed this template
-and run <code>./sbt</code>.  For example, to check out the repository and run sbt:
-
-```bash
-git clone xxxx
-cd arpegeo-template
-./sbt
-```
-
-The first time you run SBT, there will be a significant delay as all of the
-necessary libraries are downloaded.
-
-If you get any sort of error that a library can't be downloaded, please let us 
-know so we can fix the problem.  SBT downloads the libraries from where they
-are hosted on the internet, so problems can arise when repositories change.  
-
-When sbt is done loading and compiling, you'll see an sbt prompt:
-
-``` 
- >
-```
-If you type <code>run</code>, you'll see some messages from Jetty, the embedded webserver that responds to GeoTrellis requests. 
-
-```
- > run
-[info] Starting server on port 8888.
-[error] 2012-02-06 14:17:08.255:INFO:oejs.Server:jetty-8.1.0.RC4
-[error] 2012-02-06 14:17:08.329:INFO:oejs.AbstractConnector:Started SelectChannelConnector@0.0.0.0:8888 STARTING
-```
-
-This means that you are now running a webserver on port 8888 that responds to
-requests.  Go to 
-
-[http://localhost:8888/template](http://localhost:8888/template)
-
-to see the template web service, which just shows HTML of "Hello GeoTrellis!"
-
-## Step 2: Edit the configuration file
-
-There's a configuration file you can edit at src/main/resources/application.conf.  For example, some settings include:
-
-```
-geotrellis.host = "0.0.0.0" // IP address to which the server will listen
-                         // (0.0.0.0 means listen to all ports)
-geotrellis.port = 8888      // The server port 
-geotrellis.rest-package = "geotrellis.rest" // The package that will be searched for
-                                      // for rest services
-geotrellis.tmp = "/tmp"     // A temporary directory to be used for scratch files
-```
-
-## Step 3: Import data
-
-There is a script in the /scripts directory that you can use to convert raster data in
-GeoTiff format to the ARG32 format.  It assumes that you are using a Unix shell (Linux/Mac) and have installed "wget".
-
-For example, to convert ```your_geotiff.tif``` to ```output.arg32```, change to the
-directory you've checked out the template and run:
-
-```bash
-scripts/geotrellis_import_geotiff your_geotiff.tif output.arg32
-```
-
-The first time you run the script it will download the necessary GeoTrellis libraries to your
-system.
-
-## Step 4: Edit the catalog.json file
-
-The catalog.json file (in src/main/resources) defines the data sources that are
-available to the GeoTrellis service.  You can move it to an alternate location if you
-define the file path in the configuration file.
-
-## Step 5: Create your service
-
-The blank REST service is in the file src/main/scala/geotrellis/rest/TemplateResource.scala.  Edit that file to change the existing service, or create a similar 
-class in the ``geotrellis.rest`` package to create a new service.
-
-If you are using the [ScalaIDE for Eclipse](http://scala-ide.org/), you should
-type <code>eclipse</code> at the sbt prompt, and then import the project into Eclipse.
-If you are using emacs, you should consider using [ENSIME](https://github.com/aemoncannon/ensime). 
-
-## Additional notes
-
-### Triggered Restart
-
-Note: If you run ... 
-
- > ~re-start
-
-... at the sbt prompt, you will be in a 'triggered restart' mode.  As soon
-as you edit your source files, the application will re-compile and restart 
-with your updated code. 
+Look at the [template project README](https://github.com/azavea/geotrellis-template) for step by step instructions.
