+++ docs/03.md
@@ -64,6 +64,8 @@ having to handle errors at every possible point.
 Calling `server.run` returns an integer value which can be passed to the user.
 
 ```scala
+import op.util.string.ParseInt
+
 @Path("/adder/")
 class AddOne {
   @GET
@@ -119,7 +121,7 @@ class BoundingBox {
     val e2:Op[Extent] = ParseExtent(s2)
 
     // combine the extents
-    val op:Op[Extent] = CombineExtents(e1, e2)
+    val op:Op[Extent] = op.raster.extent.CombineExtents(e1, e2)
 
     // run the operation
     val data = try {
@@ -190,6 +192,8 @@ services loads and renders the entire raster at its underlying resolution,
 which can be slower.
 
 ```scala
+import geotrellis.op.util.string.{SplitOnComma,ParseInt}
+
 @Path("/draw/")
 class DrawRaster {
   @GET
@@ -200,19 +204,19 @@ class DrawRaster {
           @Context req:HttpServletRequest) = {
 
     // load the raster
-    val rasterOp = LoadRaster(name)
+    val rasterOp = raster.data.LoadRaster(name)
 
     // find the colors to use
-    val paletteOp = ForEach(SplitOnComma(palette))(ParseHexInt(_))
+    val paletteOp = logic.ForEach(SplitOnComma(palette))(ParseInt(_,16))
     val numOp = ParseInt(shades)
-    val colorsOp = ColorsFromPalette(paletteOp, numOp)
+    val colorsOp = op.stat.ColorsFromPalette(paletteOp, numOp)
 
     // find the appropriate quantile class breaks to use
-    val histogramOp = BuildMapHistogram(rasterOp)
-    val breaksOp = FindColorBreaks(histogramOp, colorsOp)
+    val histogramOp = raster.stat.Histogram(rasterOp)
+    val breaksOp = stat.ColorBreaks(histogramOp, colorsOp)
 
     // render the png
-    val pngOp = RenderPNG(rasterOp, breaksOp, 0, true)
+    val pngOp = raster.data.RenderPNG(rasterOp, breaksOp, 0, true)
 
     // run the operation
     try {
