+++ docs/05.md
@@ -1,9 +1,9 @@
 ## Operations
 
-This is a a reference table of operations you can use in your services.  A few
-caveats: this is an automatically generated table that is still in progress.  Not all of the operations are included, and a few of the return types are not 
-correct.  If you click on the names of any of the operations, you will be brought to the Scaladoc API document for the operation, which is correct.
+This wll be a reference table of operations you can use in your services, but it 
+has not yet been updated for 0.7.
 
+<!--
 <table class="bordered-table zebra-striped">
       <thead>
           <tr>
@@ -90,3 +90,4 @@ correct.  If you click on the names of any of the operations, you will be brough
 <tr><td><code><a href="http://azavea.github.com/geotrellis/latest/api/index.html#geotrellis.op.ZonalHistogram">ZonalHistogram</a></code></td><td>Histogram</td><td>Given a raster, return a histogram summary of the cells within each zone.</td></tr>
         </tbody>
       </table>
+-->
