+++ docs/08.md
@@ -6,7 +6,7 @@ that the operation acts on.  GeoTrellis will automatically run all child operati
 parallel.  For example, if you are adding each cell of five rasters together, the
 service will calculate each input raster in parallel (if resources allow).  If working
 with a single large data set (e.g. a large raster), it is usually beneficial to transform
-the raster into a ``TiledRaster`` which allows the system to parallelize the work. 
+the raster into a tiled raster which allows the system to parallelize the work. 
 
 ### Distributed Execution
 With a single command, an operation can be transformed into a RemoteOperation that
