+++ src/main/scala/geotrellis/op/raster/AsArray.scala
