+++ src/test/scala/geotrellis/data/ExtentSpec.scala
@@ -4,7 +4,8 @@ import geotrellis._
 import geotrellis.op._
 import geotrellis.process._
 import geotrellis.op.raster._
-import geotrellis.op.raster.data.{WarpRaster,LoadFile}
+import geotrellis.op.raster.transform.WarpRaster
+import geotrellis.op.io.LoadFile
 
 import org.scalatest.Spec
 import org.scalatest.matchers.MustMatchers
@@ -20,9 +21,9 @@ class ExtentSpec extends Spec with MustMatchers with ShouldMatchers {
     RasterExtent(e, cw, ch, ncols, nrows)
   }
 
-  def load(name:String) = op.raster.data.LoadRaster(name)
+  def load(name:String) = io.LoadRaster(name)
   def load(name:String, pt:(Double, Double), cs:(Double, Double), ncols:Int, nrows:Int) = {
-    op.raster.data.LoadRaster(name, re(pt, cs, ncols, nrows))
+    io.LoadRaster(name, re(pt, cs, ncols, nrows))
   }
 
   def xload(path:String) = LoadFile(path)
